pass client to waitForTransactionReceipt
